use insatnce for superstream connection
