+++ clients/src/main/java/org/apache/kafka/common/config/AbstractConfig.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.common.config;
 
+import org.apache.kafka.clients.SuperstreamConnectionHolder;
 import org.apache.kafka.common.Configurable;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.config.provider.ConfigProvider;
@@ -114,6 +115,8 @@ public class AbstractConfig {
 
         // ** Added by Superstream
         originals = Superstream.initSuperstreamConfig((Map<String, Object>) originals, type);
+        Superstream superstreamConnection = (Superstream) originals.get("superstream.connection");
+        SuperstreamConnectionHolder.initialize(superstreamConnection);
         // Added by Superstream **
         this.originals = resolveConfigVariables(configProviderProps, (Map<String, Object>) originals);
         this.values = definition.parse(this.originals);
